change bytes_written types to uint64_t

+++ include/h2o.h
@@ -1034,9 +1034,9 @@ struct st_h2o_req_t {
      */
     struct {
         struct {
-            size_t total;
-            size_t header;
-            size_t body;
+            uint64_t total;
+            uint64_t header;
+            uint64_t body;
         } bytes_written;
         h2o_httpclient_timings_t timestamps;
     } proxy_stats;
